MI_EM_amelia: Multiple imputation with EM

MI_EM_amelia is a multiple imputation function with EM algorithm. This function returns both multiple imputation results of the original incomplete data and the final imputed data (derived from multiple imputation results by Robin's Rule). As for categorical columns, both factor form and onehot probability vector form are returned.

More details about the EM implementation could be found in the documentation of function amelia from 'Amelia' package.

MI_EM_amelia(df_with_mv, col_num, col_cat = NULL, num_imp = 5)

Arguments

df_with_mv

Data matrix with missing values.

col_num

Numerical columns index

col_cat

Categorical columns index

num_imp

Number of multiple imputations

Value

ximp Final imputed data matrix. ximp.disj Final imputed data matrix of same type as 'ximp' for the numeric columns. For the categorical columns, the prediction of probability for each category is shown in form of onehot probability vector. ls_ximp List imputed data matrix from multiple imputation procedure. ls_ximp.disj List imputed data matrix from multiple imputation procedure, with categorical columns in onehot probability vector form.

References

Honaker, J., King, G., Blackwell, M. (2011). Amelia II: A Program for Missing Data. Journal of Statistical Software, 45(7), 1--47. https://www.jstatsoft.org/v45/i07/.

Statistical Analysis with Missing Data, by Little and Rubin, 2002